Feature #70119, InnerJoinController service proxy (#609)
