+++ core/src/main/java/inetsoft/web/composer/ws/event/WSJoinTablesEvent.java
@@ -17,7 +17,9 @@
  */
 package inetsoft.web.composer.ws.event;
 
-public class WSJoinTablesEvent {
+import java.io.Serializable;
+
+public class WSJoinTablesEvent implements Serializable {
    public String[] getTables() {
       return tables;
    }
